+++ lib/quality_player/src/responsive.dart
@@ -1,20 +1,16 @@
 import 'package:flutter/material.dart';
 
-class Responsive extends StatelessWidget {
-  const Responsive({super.key});
+extension Responsive on BuildContext{
+  bool isMobile() =>
+      MediaQuery.sizeOf(this).width < 850;
 
-  static bool isMobile(BuildContext context) =>
-      MediaQuery.of(context).size.width < 850;
+  bool isTablet() =>
+      MediaQuery.sizeOf(this).width < 1100 &&
+          MediaQuery.sizeOf(this).width >= 850;
 
-  static bool isTablet(BuildContext context) =>
-      MediaQuery.of(context).size.width < 1100 &&
-          MediaQuery.of(context).size.width >= 850;
+  bool isDesktop() =>
+      MediaQuery.sizeOf(this).width >= 1100;
 
-  static bool isDesktop(BuildContext context) =>
-      MediaQuery.of(context).size.width >= 1100;
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
+  double screenHeight()=>
+      MediaQuery.sizeOf(this).height;
 }